test(patina): move v-for unused-vars regression tests to own module
basic.rs already exceeds the 350-line source-length guard, and adding
the new regression tests grew the over-limit file, failing the
source-file-lengths check. Relocate the two tests into a dedicated
v_for_unused_vars module so basic.rs no longer grows.
