+++ crates/vize_patina/src/linter/tests.rs
@@ -8,3 +8,4 @@ mod jsx;
 mod jsx_fallback;
 mod no_top_level_ref;
 mod sfc;
+mod v_for_unused_vars;
